abstracted GameWorld's update()
