+++ core/src/ru/spbau/blackout/entities/ShellObject.java
@@ -3,7 +3,7 @@ package ru.spbau.blackout.entities;
 import com.badlogic.gdx.graphics.g3d.Model;
 import com.badlogic.gdx.physics.box2d.Shape;
 
-import ru.spbau.blackout.GameWorld;
+import ru.spbau.blackout.worlds.GameWorld;
 import ru.spbau.blackout.utils.Creator;
 
 public class ShellObject extends DynamicObject {
